fix: Poprawki po testach
