+++ config/routes.yaml
@@ -1,5 +1,5 @@
 controllers:
-    resource:
-        path: ../src/Controller/
-        namespace: App\Controller
+    resource: ../src/Controller/
     type: attribute
+    prefix: /
+    trailing_slash_on_root: false
