+++ src/Controller/Admin/AdminAudiobookCategoryController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Controller;
+namespace App\Controller\Admin;
 
 use App\Annotation\AuthValidation;
 use App\Entity\AudiobookCategory;
@@ -68,9 +68,10 @@ use Symfony\Contracts\Cache\TagAwareCacheInterface;
     content    : new Model(type: PermissionNotGrantedModel::class)
 )]
 #[OA\Tag(name: 'AdminAudiobookCategory')]
+#[Route('/api/admin')]
 class AdminAudiobookCategoryController extends AbstractController
 {
-    #[Route('/api/admin/category/add', name: 'adminCategoryAdd', methods: ['PUT'])]
+    #[Route('/category/add', name: 'adminCategoryAdd', methods: ['PUT'])]
     #[AuthValidation(checkAuthToken: true, roles: [UserRolesNames::ADMINISTRATOR, UserRolesNames::RECRUITER])]
     #[OA\Put(
         description: 'Endpoint is adding new category',
@@ -129,7 +130,7 @@ class AdminAudiobookCategoryController extends AbstractController
         throw new InvalidJsonDataException($translateService);
     }
 
-    #[Route('/api/admin/category/edit', name: 'adminCategoryEdit', methods: ['PATCH'])]
+    #[Route('/category/edit', name: 'adminCategoryEdit', methods: ['PATCH'])]
     #[AuthValidation(checkAuthToken: true, roles: [UserRolesNames::ADMINISTRATOR, UserRolesNames::RECRUITER])]
     #[OA\Patch(
         description: 'Endpoint is editing given category',
@@ -180,7 +181,7 @@ class AdminAudiobookCategoryController extends AbstractController
         throw new InvalidJsonDataException($translateService);
     }
 
-    #[Route('/api/admin/category/remove', name: 'adminCategoryRemove', methods: ['DELETE'])]
+    #[Route('/category/remove', name: 'adminCategoryRemove', methods: ['DELETE'])]
     #[AuthValidation(checkAuthToken: true, roles: [UserRolesNames::ADMINISTRATOR, UserRolesNames::RECRUITER])]
     #[OA\Delete(
         description: 'Endpoint is removing given category',
@@ -231,7 +232,7 @@ class AdminAudiobookCategoryController extends AbstractController
         throw new InvalidJsonDataException($translateService);
     }
 
-    #[Route('/api/admin/category/add/audiobook', name: 'adminCategoryAddAudiobook', methods: ['PUT'])]
+    #[Route('/category/add/audiobook', name: 'adminCategoryAddAudiobook', methods: ['PUT'])]
     #[AuthValidation(checkAuthToken: true, roles: [UserRolesNames::ADMINISTRATOR, UserRolesNames::RECRUITER])]
     #[OA\Put(
         description: 'Endpoint is adding audiobook to given category',
@@ -293,7 +294,7 @@ class AdminAudiobookCategoryController extends AbstractController
         throw new InvalidJsonDataException($translateService);
     }
 
-    #[Route('/api/admin/category/remove/audiobook', name: 'adminCategoryRemoveAudiobook', methods: ['DELETE'])]
+    #[Route('/category/remove/audiobook', name: 'adminCategoryRemoveAudiobook', methods: ['DELETE'])]
     #[AuthValidation(checkAuthToken: true, roles: [UserRolesNames::ADMINISTRATOR, UserRolesNames::RECRUITER])]
     #[OA\Delete(
         description: 'Endpoint is removing audiobook from given category',
@@ -356,7 +357,7 @@ class AdminAudiobookCategoryController extends AbstractController
         throw new InvalidJsonDataException($translateService);
     }
 
-    #[Route('/api/admin/category/audiobooks', name: 'adminCategoryAudiobooks', methods: ['POST'])]
+    #[Route('/category/audiobooks', name: 'adminCategoryAudiobooks', methods: ['POST'])]
     #[AuthValidation(checkAuthToken: true, roles: [UserRolesNames::ADMINISTRATOR, UserRolesNames::RECRUITER])]
     #[OA\Post(
         description: 'Endpoint is returning all audiobooks in given category',
@@ -448,7 +449,7 @@ class AdminAudiobookCategoryController extends AbstractController
         throw new InvalidJsonDataException($translateService);
     }
 
-    #[Route('/api/admin/categories/tree', name: 'adminCategoriesTree', methods: ['GET'])]
+    #[Route('/categories/tree', name: 'adminCategoriesTree', methods: ['GET'])]
     #[AuthValidation(checkAuthToken: true, roles: [UserRolesNames::ADMINISTRATOR, UserRolesNames::RECRUITER])]
     #[OA\Get(
         description: 'Endpoint is returning all categories in system as a tree',
@@ -482,7 +483,7 @@ class AdminAudiobookCategoryController extends AbstractController
         return ResponseTool::getResponse($successModel);
     }
 
-    #[Route('/api/admin/categories', name: 'adminCategories', methods: ['GET'])]
+    #[Route('/categories', name: 'adminCategories', methods: ['GET'])]
     #[AuthValidation(checkAuthToken: true, roles: [UserRolesNames::ADMINISTRATOR, UserRolesNames::RECRUITER])]
     #[OA\Get(
         description: 'Endpoint is returning all categories in system',
@@ -518,7 +519,7 @@ class AdminAudiobookCategoryController extends AbstractController
         return ResponseTool::getResponse($successModel);
     }
 
-    #[Route('/api/admin/category/active', name: 'adminCategoryActive', methods: ['PATCH'])]
+    #[Route('/category/active', name: 'adminCategoryActive', methods: ['PATCH'])]
     #[AuthValidation(checkAuthToken: true, roles: [UserRolesNames::ADMINISTRATOR, UserRolesNames::RECRUITER])]
     #[OA\Patch(
         description: 'Endpoint is activating given category',
@@ -569,7 +570,7 @@ class AdminAudiobookCategoryController extends AbstractController
         throw new InvalidJsonDataException($translateService);
     }
 
-    #[Route('/api/admin/category/detail', name: 'adminCategoryDetail', methods: ['POST'])]
+    #[Route('/category/detail', name: 'adminCategoryDetail', methods: ['POST'])]
     #[AuthValidation(checkAuthToken: true, roles: [UserRolesNames::ADMINISTRATOR, UserRolesNames::RECRUITER])]
     #[OA\Post(
         description: 'Endpoint is returning category details',
