+++ src/Controller/User/UserReportController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Controller;
+namespace App\Controller\User;
 
 use App\Annotation\AuthValidation;
 use App\Entity\Report;
@@ -55,9 +55,10 @@ use Symfony\Component\Routing\Attribute\Route;
     content    : new Model(type: PermissionNotGrantedModel::class)
 )]
 #[OA\Tag(name: 'UserReport')]
+#[Route('/api')]
 class UserReportController extends AbstractController
 {
-    #[Route('/api/report', name: 'apiReport', methods: ['PUT'])]
+    #[Route('/report', name: 'apiReport', methods: ['PUT'])]
     #[OA\Put(
         description: 'Method used to report for not logged users',
         security   : [],
@@ -137,7 +138,7 @@ class UserReportController extends AbstractController
         throw new InvalidJsonDataException($translateService);
     }
 
-    #[Route('/api/user/report', name: 'apiUserReport', methods: ['PUT'])]
+    #[Route('/user/report', name: 'apiUserReport', methods: ['PUT'])]
     #[AuthValidation(checkAuthToken: true, roles: [UserRolesNames::USER])]
     #[OA\Put(
         description: 'Endpoint is used for users to report bad behavior',
@@ -211,7 +212,7 @@ class UserReportController extends AbstractController
         throw new InvalidJsonDataException($translateService);
     }
 
-    #[Route('/api/user/reports', name: 'apiUserReports', methods: ['POST'])]
+    #[Route('/user/reports', name: 'apiUserReports', methods: ['POST'])]
     #[AuthValidation(checkAuthToken: true, roles: [UserRolesNames::USER])]
     #[OA\Post(
         description: 'Endpoint returning user reports',
